refactor: removed extra wrapper for html template parsing
